% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEAaq_get_dataframe.R
\name{EEAaq_get_dataframe}
\alias{EEAaq_get_dataframe}
\title{EEAaq_get dataframe}
\usage{
EEAaq_get_dataframe(dataframe = NULL)
}
\arguments{
\item{dataframe}{name of the dataframe to retireve}
}
\value{
a dataframe
}
\description{
Retrieve one of the datasets (LAU, NUTS, stations, or pollutant)
This function downloads and loads one dataset at a time from a predefined list of available datasets.
Ensure that the dataset name is written correctly.
}
\examples{
\donttest{
LAU <- EEAaq_get_dataframe(dataframe= "LAU")}


}
